Share Stats plan error helper
