+++ projects/packages/premium-analytics/packages/data/src/index.ts
@@ -201,8 +201,10 @@ export {
 	computeDateRangeFromPreset,
 	getApiErrorCode,
 	getApiErrorStatus,
+	getStatsPlanErrorReason,
 	shouldRetryApiError,
 } from './utils';
+export type { StatsPlanErrorReason } from './utils';
 export type { ReportDataMap } from './types';
 export type { ReportQueryParams } from './api';
 export type { FilterCondition } from './types/filter-condition';
